feat(ios): bump minimum deployment target to iOS11
and fix build error
